refactor(direct-control): extract map primitive types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only Civ7MapLocation, Civ7MapBounds, and Civ7HiddenInfoPolicy type ownership from packages/civ7-direct-control/src/index.ts into packages/civ7-direct-control/src/play/map/types.ts while keeping facade type re-exports in index.ts.

Preserves public type contracts and does not change map validation, hidden-info policy semantics, source strings, runtime behavior, procedure schemas, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- map-and-visibility.test.ts unit-move-preview.test.ts unit-target-action.test.ts tactical-reads.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.control.test.ts game.play.unit-target.test.ts game.play.tactical-read.test.ts game/play/unit-move-preview.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
