+++ packages/civ7-direct-control/src/play/map/types.ts
@@ -1,0 +1,11 @@
+export type Civ7MapLocation = Readonly<{
+  x: number;
+  y: number;
+}>;
+
+export type Civ7MapBounds = Readonly<Civ7MapLocation & {
+  width: number;
+  height: number;
+}>;
+
+export type Civ7HiddenInfoPolicy = "include-hidden" | "visibility-filtered" | "not-player-scoped";
